+++ graph/graph.go
@@ -5,6 +5,7 @@ package graph
 import (
 	"bytes"
 	"fmt"
+	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/regen-network/regen-ledger/types"
 	"golang.org/x/crypto/blake2b"
 	"net/url"
@@ -78,6 +79,39 @@ type Class interface {
 	// RequiredProperties returns a slice of properties with arity one that are required
 	RequiredProperties() []Property
 	fmt.Stringer
+}
+
+// AccAddressID wraps sdk.AccAddress as a Node ID type
+type AccAddressID struct {
+	sdk.AccAddress
+}
+
+// URI returns the URI representation of the AccAddressID
+func (id AccAddressID) URI() *url.URL {
+	uri, err := url.Parse(id.String())
+	if err != nil {
+		panic(err)
+	}
+	return uri
+}
+
+// HashID represents a node ID that is a URI hash fragment of the unlabelled root node
+type HashID struct {
+	Fragment string
+}
+
+func (id HashID) String() string {
+	return fmt.Sprintf("_:_.%s", id.Fragment)
+}
+
+// URI returns the URI representation of the HashID
+func (id HashID) URI() *url.URL {
+	// TODO this should really be a blank node or the actual hash of the graph plus the fragment
+	uri, err := url.Parse(fmt.Sprintf("root:#%s", id.Fragment))
+	if err != nil {
+		panic(err)
+	}
+	return uri
 }
 
 func (ty PropertyType) String() string {
